+++ app/controllers/ae_module/adverse_events_controller.rb
@@ -4,7 +4,9 @@
 class AeModule::AdverseEventsController < AeModule::BaseController
   before_action :find_project_as_reporter_or_admin_or_team_member_or_redirect
   before_action :redirect_blinded_users
-  before_action :find_adverse_event_or_redirect, only: [:show, :edit, :update, :log]
+  before_action :find_adverse_event_or_redirect, only: [
+    :show, :edit, :update, :history, :log
+  ]
   before_action :set_project_member
   layout :sidebar_layout
 
@@ -16,6 +18,12 @@ class AeModule::AdverseEventsController < AeModule::BaseController
   # # GET /projects/:project_id/ae-module/adverse-events/:id
   # def show
   # end
+
+  # GET /projects/:project_id/ae-module/adverse-events/:id/history
+  def history
+    @viewable_adverse_event_ids = adverse_events.pluck(:id)
+    @adverse_events = @adverse_event.historical_events.page(params[:page]).per(20)
+  end
 
   # # GET /projects/:project_id/ae-module/adverse-events/:id/log
   # def log