+++ app/models/ae_adverse_event.rb
@@ -62,6 +62,10 @@ class AeAdverseEvent < ApplicationRecord
     self.class.where(project: project).order(:created_at).pluck(:id).index(id)&.send(:+, 1)
   end
 
+  def historical_events
+    subject.ae_adverse_events
+  end
+
   def sent_for_review?
     !sent_for_review_at.nil?
   end